remove ExceptionController and refactor ErrorController
