+++ packages/framework/src/Resources/config/routing.yaml
@@ -2,3 +2,18 @@ shopsys_framework_admin:
     resource: "@ShopsysFrameworkBundle/Controller"
     prefix: /admin
     type: annotation
+
+front_error_page:
+    path: /_error/{code}/
+    defaults:
+        _controller: Shopsys\FrameworkBundle\Component\Error\ErrorController:errorPagePreviewAction
+    requirements:
+        code: \d+
+
+front_error_page_format:
+    path: /_error/{code}/{_format}/
+    defaults:
+        _controller: Shopsys\FrameworkBundle\Component\Error\ErrorController:errorPagePreviewAction
+    requirements:
+        code: \d+
+        _format: css|html|js|json|txt|xml
