+++ project-base/config/packages/framework.yaml
@@ -21,4 +21,4 @@ framework:
         enabled: false
         collect: false
     assets: ~
-    error_controller: 'App\Controller\Front\ErrorController::showAction'
+    error_controller: 'Shopsys\FrameworkBundle\Component\Error\ErrorController::showAction'
