+++ project-base/config/routes/shopsys_front.yaml
@@ -48,21 +48,6 @@ front_customer_login_as_remembered_user:
     defaults:
         _controller: App\Controller\Front\CustomerController:loginAsRememberedUserAction
 
-front_error_page:
-    path: /_error/{code}/
-    defaults:
-        _controller: App\Controller\Front\ErrorController:errorPageAction
-    requirements:
-        code: \d+
-
-front_error_page_format:
-    path: /_error/{code}/{_format}/
-    defaults:
-        _controller: App\Controller\Front\ErrorController:errorPageAction
-    requirements:
-        code: \d+
-        _format: css|html|js|json|txt|xml
-
 front_homepage:
     path: /
     defaults:
